Tonne of changes relating to ball collisions. Super buggy atm.

+++ prototypes/collisions/src/collisionsModel/LeftFlipper.java
@@ -2,7 +2,7 @@ package collisionsModel;
 
 public class LeftFlipper extends Gizmo {
 
-    public LeftFlipper(String id,double x, double y) {
+    public LeftFlipper(String id, int x, int y) {
         super(id,x,y);
     }
 }
