+++ prototypes/collisions/src/collisionsModel/RightFlipper.java
@@ -2,7 +2,7 @@ package collisionsModel;
 
 public class RightFlipper extends Gizmo {
 
-    public RightFlipper(String id,double x, double y) {
+    public RightFlipper(String id, int x, int y) {
         super(id,x,y);
     }
 }
